cystring: Rewrite compose_str to grow a buffer as needed to write a string
The original implementation of compose_str used vsnprintf to measure the
size of the formatted string prior to allocation. This can be slow since
we will always have to "write the string twice". For small strings, a
sufficiently large initial buffer size will allow us to get away with
only writing once.